Fix the title bars
