+++ tools/PI/DevHome.PI/BarWindowVertical.xaml
@@ -1,0 +1,122 @@
+<?xml version="1.0" encoding="utf-8"?>
+<winex:WindowEx
+    x:Class="DevHome.PI.BarWindowVertical"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:winex="using:WinUIEx"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:controls="using:DevHome.PI.Controls"
+    xmlns:helpers="using:DevHome.PI.Helpers"
+    mc:Ignorable="d"
+    Title="" MinHeight="700" MinWidth="70" MaxWidth="70" Width="70" Height="700"
+    TaskBarIcon="Images/pi.ico" IsTitleBarVisible="False"
+    Closed="WindowEx_Closed" >
+
+    <!-- TODO Make DesktopAcrylicBackdrop/MicaBackdrop and AcrylicBackgroundFillColorBaseBrush user-configurable.-->
+    <Window.SystemBackdrop>
+        <MicaBackdrop/>
+    </Window.SystemBackdrop>
+
+    <Grid x:Name="MainPanel" Loaded="MainPanel_Loaded" HorizontalAlignment="Center">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <!-- Chrome buttons for a vertical bar -->
+        <StackPanel x:Name="ChromeButtonPanelVertical" Orientation="Vertical" HorizontalAlignment="Center" Margin="0, 10" Grid.Row="0">
+            <TextBlock x:Name="GripTextBlock" Text="&#xe76f;&#xe76f;" FontSize="10" Margin="0" HorizontalAlignment="Center" FontFamily="Segoe Fluent Icons"  PointerPressed="Window_PointerPressed" PointerMoved="Window_PointerMoved" PointerReleased="Window_PointerReleased"/>
+                <Button
+                    Margin="0,20, 0, 0"
+                    Style="{StaticResource ChromeButton}"
+                    HorizontalAlignment="Center"
+                    IsEnabled="{x:Bind viewModel.IsSnappingEnabled, Mode=OneWay}"
+                    Click="{x:Bind viewModel.PerformSnapCommand}" >
+                <TextBlock Text="{x:Bind viewModel.CurrentSnapButtonText, Mode=OneWay}"/>
+                </Button>
+                <Button
+                    Style="{StaticResource ChromeButton}"
+                    Click="CloseButton_Click" 
+                    HorizontalAlignment="Center"
+                    Margin="0,20, 0, 0" >
+                    <TextBlock Text="&#xE8BB;"/>
+                </Button>
+        </StackPanel>
+
+
+        <StackPanel x:Name="AllControls" Orientation="Vertical" Grid.Row="1" HorizontalAlignment="Center" Margin="0" Background="Transparent">
+            <StackPanel.Resources>
+                <Style TargetType="TextBlock">
+                    <Setter Property="FontFamily" Value="Segoe Fluent Icons"/>
+                    <Setter Property="FontSize" Value="20"/>
+                </Style>
+                <Style TargetType="Button">
+                    <Setter Property="BorderThickness" Value="0"/>
+                    <Setter Property="Background" Value="Transparent"/>
+                    <Setter Property="HorizontalAlignment" Value="Center"/>
+                </Style>
+                <Style TargetType="controls:ProcessSelectionButton">
+                    <Setter Property="BorderThickness" Value="0"/>
+                    <Setter Property="Background" Value="Transparent"/>
+                </Style>
+            </StackPanel.Resources>
+                        
+            <controls:ProcessSelectionButton x:Uid="ProcessChooserButton" x:Name="ProcessChooserButton" Click="{x:Bind viewModel.ProcessChooserCommand}" HorizontalAlignment="Center" Margin="0">
+                <TextBlock Text="&#xecaa;" Margin="0" />
+            </controls:ProcessSelectionButton>
+
+            <!-- Per App controls -->
+            <Image x:Name="AppIcon" HorizontalAlignment="Center" Source="{x:Bind viewModel.ApplicationIcon, Mode=OneWay}" Margin="5" Height="20" Width="20" Visibility="{x:Bind viewModel.AppBarVisibility, Mode=OneWay}"/>
+            <TextBlock x:Uid="AppPID" Text="{x:Bind viewModel.ApplicationPid, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" Margin="5" HorizontalAlignment="Center" Visibility="{x:Bind viewModel.AppBarVisibility, Mode=OneWay}"/>
+            <TextBlock x:Uid="AppCPUUsage" Text="{x:Bind viewModel.AppCpuUsage, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" Margin="5" HorizontalAlignment="Center" Visibility="{x:Bind viewModel.AppBarVisibility, Mode=OneWay}"/>
+           
+            <!-- Per System controls -->
+            <Button x:Uid="SwitchLayoutButton" x:Name="SwitchLayoutButton" Click="{x:Bind viewModel.SwitchLayoutCommand}" HorizontalAlignment="Center">
+                <TextBlock Text="&#xe8b4;"/>
+            </Button>
+            <Button x:Name="LargeContentButton" Click="{x:Bind viewModel.ShowBigWindowCommand}" HorizontalAlignment="Center">
+                <TextBlock Text="&#xe8a1;"/>
+            </Button>
+
+            <!-- TODO When we add a tool, we should add it to the ExternalToolsMenu. -->
+            <Button x:Name="ExternalToolsButton" x:Uid="ExternalToolsButton" HorizontalAlignment="Center">
+                <Button.ContextFlyout>
+                    <MenuFlyout x:Name="ExternalToolsMenu">
+                        <!-- Menuitems for each external tool will be added below. -->
+                    </MenuFlyout>
+                </Button.ContextFlyout>
+                <TextBlock Text="&#xEC7A;"/>
+            </Button>
+
+            <ItemsRepeater x:Name="ExternalToolsRepeater" ItemsSource="{x:Bind helpers:ExternalToolsHelper.Instance.ExternalTools, Mode=OneWay}" Layout="{StaticResource ExternalToolsVerticalLayout}">
+                <ItemsRepeater.ContextFlyout>
+                    <MenuFlyout x:Name="ToolContextMenu">
+                        <MenuFlyoutItem x:Uid="UnPinMenuItem" x:Name="UnPinMenuItem" Click="UnPinMenuItem_Click">
+                            <MenuFlyoutItem.Icon>
+                                <FontIcon Glyph="&#xE77A;"/>
+                            </MenuFlyoutItem.Icon>
+                        </MenuFlyoutItem>
+                        <MenuFlyoutItem x:Uid="UnregisterMenuItem" x:Name="UnregisterMenuItem" Click="UnregisterMenuItem_Click">
+                            <MenuFlyoutItem.Icon>
+                                <FontIcon Glyph="&#xECC9;"/>
+                            </MenuFlyoutItem.Icon>
+                        </MenuFlyoutItem>
+                    </MenuFlyout>
+                </ItemsRepeater.ContextFlyout>
+                <DataTemplate x:DataType="helpers:ExternalTool">
+                    <Button Click="ExternalToolButton_Click" HorizontalAlignment="Center" Width="45" Height="45" ToolTipService.ToolTip="{x:Bind Name, Mode=OneWay}" Tag="{x:Bind}" PointerPressed="ExternalToolButton_PointerPressed">
+                        <Image Source="{x:Bind ToolIcon, Mode=OneWay}"/>
+                    </Button>
+                </DataTemplate>
+            </ItemsRepeater>
+        </StackPanel>
+
+        <StackPanel x:Name="SystemResourceStackPanel" Orientation="Vertical" Grid.Row="2" Margin="0,0,0,5">
+            <TextBlock x:Uid="SystemCPUUsage" Text="{x:Bind viewModel.SystemCpuUsage, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5, 0, 5, 0"/>
+            <TextBlock x:Uid="SystemMemUsage" Text="{x:Bind viewModel.SystemRamUsage, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5, 5, 5, 5"/>
+            <TextBlock x:Uid="SystemDiskUsage" Text="{x:Bind viewModel.SystemDiskUsage, Mode=OneWay}"  FontFamily="Segoe UI" FontSize="10" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="5, 0, 5, 0"/>
+        </StackPanel>
+    </Grid>
+</winex:WindowEx>
