--- tools/PI/DevHome.PI/Controls/AppStatusBar.xaml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<UserControl
-    x:Class="DevHome.PI.Controls.AppStatusBar"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <StackPanel x:Name="AppItems" Orientation="{x:Bind viewModel.BarWindow.BarOrientation, Mode=OneWay}" HorizontalAlignment="Center" Margin="0" Visibility="{x:Bind viewModel.AppBarVisibility, Mode=OneWay}">
-        <Image x:Name="AppIcon" VerticalAlignment="Center" HorizontalAlignment="Center" Source="{x:Bind viewModel.ApplicationIcon, Mode=OneWay}" Height="20" Width="20"/>
-        <TextBlock x:Uid="AppPID" Text="{x:Bind viewModel.ApplicationPid, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" VerticalAlignment="Center" Margin="5,5,5,5" HorizontalAlignment="Center"/>
-        <TextBlock x:Uid="AppCPUUsage" Text="{x:Bind viewModel.CpuUsage, Mode=OneWay}" FontFamily="Segoe UI" FontSize="10" VerticalAlignment="Center" HorizontalAlignment="Center"/>
-    </StackPanel>
-</UserControl>
