--- tools/PI/DevHome.PI/Controls/SystemBarHorizontal.xaml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<local:SystemBar
-    x:Class="DevHome.PI.Controls.SystemBarHorizontal"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:DevHome.PI.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid x:Name="HorizontalBar">
-        <TextBlock x:Name="TitleTextBlock" Text="DevHome PI" FontSize="10" Margin="10,5,0,0" HorizontalAlignment="Left" />
-        <TextBlock x:Name="GripTextBlock" Text="&#xe76f;&#xe76f;&#xe76f;&#xe76f;" FontSize="10" Margin="0,5,0,0" HorizontalAlignment="Center" FontFamily="Segoe Fluent Icons"/>
-        <StackPanel x:Name="ChromeButtonPanel" Orientation="Horizontal" HorizontalAlignment="Right" Margin="0, 5, 10, 0" >
-            <Button
-             x:Uid="SnapButton"
-             x:Name="SnapButton" Style="{StaticResource ChromeButton}"
-             Click="SnapButton_Click"
-             IsEnabled="{x:Bind IsSnappingEnabled, Mode=OneWay}">
-                <TextBlock Text="{x:Bind CurrentSnapButtonText, Mode=OneWay}"/>
-            </Button>
-            <Button
-         x:Uid="MinimizeButton"
-         x:Name="MinimizeButton" Style="{StaticResource ChromeButton}"
-         Click="MinimizeButton_Click" Margin="20,0,0,0">
-                <TextBlock Text="&#xE921;"/>
-            </Button>
-            <Button
-         x:Uid="MaximizeButton"
-         x:Name="MaximizeButton" Style="{StaticResource ChromeButton}"
-         Click="MaximizeButton_Click" Margin="20,0,0,0" Visibility="{x:Bind MaximizeButtonVisibility, Mode=OneWay}">
-                <TextBlock Text="&#xE922;"/>
-            </Button>
-            <Button
-         x:Uid="RestoreButton"
-         x:Name="RestoreButton" Style="{StaticResource ChromeButton}"
-         Click="RestoreButton_Click" Margin="20,0,0,0" Visibility="{x:Bind RestoreButtonVisibility, Mode=OneWay}">
-                <TextBlock Text="&#xE923;"/>
-            </Button>
-            <Button
-         x:Uid="CloseButton"
-         x:Name="CloseButton" Style="{StaticResource ChromeButton}"
-         Click="CloseButton_Click" Margin="20,0,0,0" >
-                <Button.ContextFlyout>
-                    <MenuFlyout x:Name="TopMenu">
-                        <MenuFlyoutItem
-                     x:Uid= "CloseAllMenuItem"
-                     x:Name="CloseAllMenuItem"
-                     Click="CloseAllMenuItem_Click">
-                            <MenuFlyoutItem.Icon>
-                                <FontIcon Glyph="&#xE711;"/>
-                            </MenuFlyoutItem.Icon>
-                        </MenuFlyoutItem>
-                    </MenuFlyout>
-                </Button.ContextFlyout>
-                <TextBlock Text="&#xE8BB;"/>
-            </Button>
-        </StackPanel>
-    </Grid>
-</local:SystemBar>
