--- tools/PI/DevHome.PI/Controls/SystemBarHorizontal.xaml.cs
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace DevHome.PI.Controls;
-
-// This is the class for a horiztonal system/chrome/title bar.
-public sealed partial class SystemBarHorizontal : SystemBar
-{
-    public SystemBarHorizontal()
-    {
-        this.InitializeComponent();
-    }
-}
