--- tools/PI/DevHome.PI/Controls/SystemBarVertical.xaml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<local:SystemBar
-    x:Class="DevHome.PI.Controls.SystemBarVertical"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:DevHome.PI.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid x:Name="VerticalBar">
-        <StackPanel x:Name="ChromeButtonPanelVertical" Orientation="Vertical" HorizontalAlignment="Center" Margin="6, 10, 6, 10">
-            <Button
-            Style="{StaticResource ChromeButton}"
-            Click="SnapButton_Click" >
-                <TextBlock Text="{x:Bind CurrentSnapButtonText, Mode=OneWay}"/>
-            </Button>
-            <Button
-        Style="{StaticResource ChromeButton}"
-        Click="CloseButton_Click" Margin="0,20, 0, 0" >
-                <Button.ContextFlyout>
-                    <MenuFlyout>
-                        <MenuFlyoutItem
-                    Click="CloseAllMenuItem_Click">
-                            <MenuFlyoutItem.Icon>
-                                <FontIcon Glyph="&#xE711;"/>
-                            </MenuFlyoutItem.Icon>
-                        </MenuFlyoutItem>
-                    </MenuFlyout>
-                </Button.ContextFlyout>
-                <TextBlock Text="&#xE8BB;"/>
-            </Button>
-
-        </StackPanel>
-    </Grid>
-</local:SystemBar>
