--- tools/PI/DevHome.PI/Controls/SystemBarVertical.xaml.cs
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace DevHome.PI.Controls;
-
-// This is the class for a vertical system/chrome/title bar.
-public sealed partial class SystemBarVertical : SystemBar
-{
-    public SystemBarVertical()
-    {
-        this.InitializeComponent();
-    }
-}
