+++ tools/PI/DevHome.PI/DevHome.PI.csproj
@@ -27,11 +27,8 @@
 
   <ItemGroup>
     <None Remove="BarWindow.xaml" />
-    <None Remove="Controls\AppStatusBar.xaml" />
     <None Remove="Controls\ExpandedViewControl.xaml" />
     <None Remove="Controls\GlowButton.xaml" />
-    <None Remove="Controls\SystemBarHorizontal.xaml" />
-    <None Remove="Controls\SystemBarVertical.xaml" />
     <None Remove="errors.db" />
     <None Remove="Images\PI.ico" />
     <None Remove="Pages\AppDetailsPage.xaml" />
@@ -77,7 +74,7 @@
     <PackageReference Include="CommunityToolkit.WinUI.Controls.Primitives" Version="8.0.240109" />
     <PackageReference Include="CommunityToolkit.WinUI.Controls.Sizers" Version="8.0.240109" />
     <PackageReference Include="Microsoft.Data.Sqlite" Version="8.0.3" />
-    <PackageReference Include="Microsoft.Diagnostics.Tracing.TraceEvent" Version="3.1.9" GeneratePathProperty="true"/>
+    <PackageReference Include="Microsoft.Diagnostics.Tracing.TraceEvent" Version="3.1.9" GeneratePathProperty="true" />
     <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="8.0.0" />
     <PackageReference Include="Microsoft.Windows.CsWin32" Version="0.3.49-beta">
       <PrivateAssets>all</PrivateAssets>
@@ -150,15 +147,6 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Update="Pages\InsightsPage.xaml">
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Update="Controls\SystemBarHorizontal.xaml">
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Update="Controls\SystemBarVertical.xaml">
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Update="Controls\AppStatusBar.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <None Update="appsettings_pi.json">
