+++ tools/PI/DevHome.PI/Program.cs
@@ -239,7 +239,7 @@ public static class Program
                             barWindow.NavigateTo(pageType);
                         }
 
-                        barWindow.ShouldExpandLargeWindow = true;
+                        // barWindow.ShouldExpandLargeWindow = true;
                     }
                 });
             }
