--- tools/PI/DevHome.PI/ViewModels/AppStatusBarViewModel.cs
@@ -1,95 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System.ComponentModel;
-using CommunityToolkit.Mvvm.ComponentModel;
-using DevHome.PI.Helpers;
-using DevHome.PI.Models;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Media.Imaging;
-
-namespace DevHome.PI.ViewModels;
-
-public partial class AppStatusBarViewModel : ObservableObject
-{
-    private readonly Microsoft.UI.Dispatching.DispatcherQueue dispatcher;
-
-    [ObservableProperty]
-    private string cpuUsage = string.Empty;
-
-    [ObservableProperty]
-    private Visibility appBarVisibility = Visibility.Visible;
-
-    [ObservableProperty]
-    private int applicationPid;
-
-    [ObservableProperty]
-    private SoftwareBitmapSource? applicationIcon;
-
-    public BarWindow? BarWindow { get; set; }
-
-    public AppStatusBarViewModel()
-    {
-        dispatcher = Microsoft.UI.Dispatching.DispatcherQueue.GetForCurrentThread();
-        PerfCounters.Instance.PropertyChanged += PerfCounterHelper_PropertyChanged;
-        TargetAppData.Instance.PropertyChanged += TargetApp_PropertyChanged;
-
-        var process = TargetAppData.Instance.TargetProcess;
-
-        AppBarVisibility = process is null ? Visibility.Collapsed : Visibility.Visible;
-
-        if (process != null)
-        {
-            ApplicationPid = process.Id;
-            ApplicationIcon = TargetAppData.Instance.Icon;
-        }
-    }
-
-    internal void SetBarWindow(BarWindow barWindow)
-    {
-        BarWindow = barWindow;
-    }
-
-    private void PerfCounterHelper_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-    {
-        if (e.PropertyName == nameof(PerfCounters.CpuUsage))
-        {
-            dispatcher.TryEnqueue(() =>
-            {
-                CpuUsage = CommonHelper.GetLocalizedString("CpuPerfTextFormatNoLabel", PerfCounters.Instance.CpuUsage);
-            });
-        }
-    }
-
-    private void TargetApp_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-    {
-        if (e.PropertyName == nameof(TargetAppData.TargetProcess))
-        {
-            var process = TargetAppData.Instance.TargetProcess;
-
-            dispatcher.TryEnqueue(() =>
-            {
-                // The App status bar is only visibile if we're attached to a process
-                AppBarVisibility = process is null ? Visibility.Collapsed : Visibility.Visible;
-
-                if (process is not null)
-                {
-                    ApplicationPid = process.Id;
-                }
-            });
-        }
-        else if (e.PropertyName == nameof(TargetAppData.Icon))
-        {
-            SoftwareBitmapSource? icon = TargetAppData.Instance?.Icon;
-
-            dispatcher.TryEnqueue(() =>
-            {
-                ApplicationIcon = icon;
-            });
-        }
-        else if (e.PropertyName == nameof(TargetAppData.HasExited))
-        {
-            // TODO Grey ourselves out if the app has exited?
-        }
-    }
-}
